refactor: consolidate useAdultVerification tests and polish hlsAuthLoader doc
